Fix issues with fillmodule
In F43 the activity failed to build and this was because
there was an int and unsigned long division in eggfill.c
which threw an error, this commit fixes that.

This commit also fixes the initialization of fillmodule,
a leftover from the Python3 port as the old initialization
was for Python2 but deprecated for Python3.

I've also changed the compile instructions to include testing
outside the activity directory.

Signed-off-by: Chihurumnaya Ibiam <ibiam@sugarlabs.org>